[lldb][test] Combine libstdc++ and libc++ tuple tests into generic test
This combines the libc++ and libstdc++ test cases. The main difference
was that the libstdcpp tests had some tuple indexing tests that libc++ didn't have.

Split out from https://github.com/llvm/llvm-project/pull/146740
