--- lldb/test/API/functionalities/data-formatter/data-formatter-stl/libstdcpp/tuple/Makefile
@@ -1,5 +1,0 @@
-CXX_SOURCES := main.cpp
-
-USE_LIBSTDCPP := 1
-
-include Makefile.rules
